+++ include/kernaux/console.h
@@ -5,12 +5,14 @@
 extern "C" {
 #endif
 
+#include <stddef.h>
+
 void kernaux_console_print(const char *s);
 void kernaux_console_printf(const char *format, ...)
 __attribute__((format(printf, 1, 2)));
 void kernaux_console_putc(char c);
 void kernaux_console_puts(const char *s);
-void kernaux_console_write(const char *data, unsigned int size);
+void kernaux_console_write(const char *data, size_t size);
 
 #ifdef __cplusplus
 }
